Initialize redux store
