+++ src/components/Todos.jsx
@@ -1,28 +1,10 @@
 import React from 'react';
 import { Todo } from './Todo';
 import { List } from '@mui/material';
+import { useSelector } from 'react-redux';
 
 export const Todos = () => {
-	const todos = [
-		{
-			userId: 1,
-			id: 1,
-			title: 'delectus aut autem',
-			completed: false
-		},
-		{
-			userId: 1,
-			id: 2,
-			title: 'quis ut nam facilis et officia qui',
-			completed: false
-		},
-		{
-			userId: 1,
-			id: 3,
-			title: 'fugiat veniam minus',
-			completed: false
-		}
-	];
+	const todos = useSelector(state => state.todos);
 	return (
 		<List sx={{ width: '100%' }}>
 			{todos.map(props => (
