+++ src/reducers/index.js
@@ -1,0 +1,5 @@
+import todoReducer from './todoReducer';
+import { combineReducers } from 'redux';
+
+const allReducer = combineReducers({ todos: todoReducer });
+export default allReducer;
